Stream output to buffer asynchronously  (#23)
Until now, when the command was executed, we would wait until the
program exists, to print the output to the buffer. So, after executing a
slightly long-running command, you'd have to wait for an awkward few
seconds until you saw any output.

This commit changes that by streaming the output to the buffer as and
when something is written to STDOUT. This way, the user will see
some kind of progress while the command is executing.

This was possible with the help of Neovim's jobstart() function. This
lets us run the command asynchronously and sends us callback signals
when something is written to STDOUT, STDERR, or the command exits.