Edit custom commands in a floating window
Previously, the edit window for custom commands was spawned using
"10split". This would create a small split buffer. But this forced us to
use global autocommands to clean up this window/buffer on save and exit.

This commit changes that by taking a more modern approach by spawning a
floating window for editing custom commands. The clean up is done using
a buffer local autocommand

Closes #31